Refactor endpoints in own files
