+++ src/api/__init__.py
@@ -1,1 +1,2 @@
 from src.api.ipfs import ipfs_bp
+from src.api.pql import pql_bp
